Add files via upload
These are custom css changes added to the original so that it's not fully blacked out. 
For the color you would like, please rename the file to "custom.css" and save inside of the darkblend folder.
Once you do so, refresh your whmcs page to see changes take into effect.